+++ components/sections/index.ts
@@ -1,5 +1,11 @@
 // HatchingCoreHeroSection removed - replaced by DynamicHeroSection
 export { default as DarkFeatureSection } from './DarkFeatureSection'
+export { default as FounderStatementSection } from './FounderStatementSection'
 export { default as LightFeatureSection } from './LightFeatureSection'
 export { default as PhilosophySection } from './PhilosophySection'
-export { default as FounderStatementSection } from './FounderStatementSection'
+
+// Product Page Sections
+export { default as FeatureBreakdownSection } from './FeatureBreakdownSection'
+export { default as FinalCTASection } from './FinalCTASection'
+export { default as PainSolutionSection } from './PainSolutionSection'
+export { default as PricingSection } from './PricingSection'